starter_code/backend/requirements.txt
flask
Flask-SQLAlchemy
flask-cors
python-dotenv
python-jose
six
sqlalchemy